+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -23,6 +23,9 @@ exclude: true
 - Added overloads of [`GeckoSession.loadUri()`][71.6] that accept a map of 
   additional HTTP request headers.
 - Added support for exposing the content blocking log in [`ContentBlockingController`][71.7].({{bugzilla}}1580201)
+- ⚠️  Added `nativeApp` to [`WebExtension.MessageDelegate.onMessage`][71.8] which
+  exposes the native application identifier that was used to send the message.
+  ({{bugzilla}}1546445)
 
 [71.1]: {{javadoc_uri}}/RuntimeTelemetry.Delegate.html#onBooleanScalar-org.mozilla.geckoview.RuntimeTelemetry.Metric-
 [71.2]: {{javadoc_uri}}/RuntimeTelemetry.Delegate.html#onLongScalar-org.mozilla.geckoview.RuntimeTelemetry.Metric-
@@ -31,6 +34,7 @@ exclude: true
 [71.5]: {{javadoc_uri}}/RuntimeTelemetry.Metric.html
 [71.6]: {{javadoc_uri}}/GeckoSession.html#loadUri-java.lang.String-java.io.File-java.util.Map-
 [71.7]: {{javadoc_uri}}/ContentBlockingController.html
+[71.8]: {{javadoc_uri}}/WebExtension.MessageDelegate.html#onMessage-java.lang.String-java.lang.Object-org.mozilla.geckoview.WebExtension.MessageSender-
 
 ## v70
 - Added API for session context assignment
@@ -353,4 +357,4 @@ exclude: true
 [65.24]: {{javadoc_uri}}/CrashReporter.html#sendCrashReport-android.content.Context-android.os.Bundle-java.lang.String-
 [65.25]: {{javadoc_uri}}/GeckoResult.html
 
-[api-version]: 6e353ab446ba296fd348d123651d024d21a15651
+[api-version]: 9fd93f90a9043b290ca9a3387228f00adfcedb51